fix: move interface to its own type file
