+++ source/services/monitoring/MonitoringService.types.ts
@@ -1,0 +1,7 @@
+import { ComponentType } from "react";
+
+export interface MonitoringService {
+  init: () => void;
+  sendError: (Error: Error | string | null) => void;
+  wrap: (component: ComponentType<unknown>) => ComponentType<unknown>;
+}
